Phase 2 step 8: Extract the 7 modal/widget classes to screens/ + widgets/ (fix load_tms_settings stub)
