+++ apps/tui/dashboard_tui.py
@@ -471,904 +471,25 @@ _stats_io._account_initial_capital_from_files = _persistence._account_initial_ca
 # WIDGETS
 # ═══════════════════════════════════════════════════════════════════════════════
 
-class MarketPanel(Widget):
-    """Titled panel with a DataTable inside."""
-
-    def __init__(self, panel_title: str = "", title_color: str = AMBER, **kw):
-        super().__init__(**kw)
-        self._panel_title = panel_title
-        self._title_color = title_color
-
-    def compose(self) -> ComposeResult:
-        yield Static(self._panel_title, classes="panel-title")
-        yield DataTable(zebra_stripes=True, cursor_type="row")
-
-    def set_data(self, columns: list[tuple[str, str]], rows: list[list]):
-        dt = self.query_one(DataTable)
-        dt.clear(columns=True)
-        for label, key in columns:
-            dt.add_column(label, key=key)
-        for row in rows:
-            dt.add_row(*row)
-
-    def update_title(self, title: str):
-        self.query_one(".panel-title", Static).update(title)
+from apps.tui.widgets.market_panel import MarketPanel  # re-export
 
 
 # ═══════════════════════════════════════════════════════════════════════════════
 # MODAL DIALOGS
 # ═══════════════════════════════════════════════════════════════════════════════
 
-class ModalDialog(ModalScreen[dict | None]):
-    """Buy/Sell/Lookup modal dialog."""
-
-    DEFAULT_CSS = """
-    ModalDialog { align: center middle; }
-    """
-
-    def __init__(
-        self,
-        mode: str = "buy",
-        *,
-        initial_symbol: str = "",
-        initial_shares: str = "",
-        initial_price: str = "",
-        default_slippage: str = "2.0",
-        holdings_positions: list[dict] | None = None,
-    ):
-        super().__init__()
-        self.mode = mode
-        self.initial_symbol = initial_symbol
-        self.initial_shares = initial_shares
-        self.initial_price = initial_price
-        self.default_slippage = default_slippage
-        self.holdings_summary = _format_sell_holdings_summary(holdings_positions)
-
-    def compose(self) -> ComposeResult:
-        with Vertical(id="dialog-box", classes=f"dialog-shell dialog-{self.mode}"):
-            title_map = {"buy": "BUY", "sell": "SELL", "lookup": "LOOKUP"}
-            yield Static(title_map.get(self.mode, ""), classes="dialog-title")
-            if self.mode == "buy":
-                yield Static("Create a paper ticket for the daily order book.", id="dialog-kicker")
-            elif self.mode == "sell":
-                yield Static("Holdings available to sell", id="dialog-kicker")
-                yield Static(self.holdings_summary, id="dialog-holdings")
-            yield Input(id="inp-symbol", placeholder="Symbol e.g. NABIL")
-            if self.mode != "lookup":
-                yield Input(id="inp-shares",
-                            placeholder="Shares" if self.mode == "buy" else "Shares (or all)")
-                yield Input(id="inp-price", placeholder="Price (blank=last close)")
-                yield Input(id="inp-slippage", placeholder="Slippage %", value=self.default_slippage)
-            yield Static("", id="dialog-result")
-            with Horizontal(id="dialog-buttons"):
-                yield Button("Submit", id="btn-submit", variant="success")
-                yield Button("Cancel", id="btn-cancel", variant="error")
-
-    def on_mount(self) -> None:
-        symbol = self.query_one("#inp-symbol", Input)
-        symbol.value = self.initial_symbol
-        if self.mode != "lookup":
-            self.query_one("#inp-shares", Input).value = self.initial_shares
-            self.query_one("#inp-price", Input).value = self.initial_price
-        symbol.focus()
-
-    def on_input_submitted(self, event: Input.Submitted) -> None:
-        self._submit()
-
-    def on_button_pressed(self, event: Button.Pressed) -> None:
-        if event.button.id == "btn-submit":
-            self._submit()
-        else:
-            self.dismiss(None)
-
-    def _submit(self):
-        sym = self.query_one("#inp-symbol", Input).value.strip().upper()
-        if not sym:
-            self.query_one("#dialog-result", Static).update(
-                Text("Symbol required", style=LOSS_HI))
-            return
-        if self.mode == "lookup":
-            self.dismiss({"symbol": sym})
-        else:
-            shares = self.query_one("#inp-shares", Input).value.strip()
-            price = self.query_one("#inp-price", Input).value.strip()
-            slippage = self.query_one("#inp-slippage", Input).value.strip()
-            if not shares:
-                self.query_one("#dialog-result", Static).update(
-                    Text("Shares required", style=LOSS_HI))
-                return
-            self.dismiss({"symbol": sym, "shares": shares, "price": price, "slippage": slippage})
-
-    def key_escape(self) -> None:
-        self.dismiss(None)
+from apps.tui.screens.dialog import ModalDialog  # re-export
 
 
 # ═══════════════════════════════════════════════════════════════════════════════
 # MODE SELECT SCREEN — shown on startup
 # ═══════════════════════════════════════════════════════════════════════════════
 
-
-class ModeSelectScreen(ModalScreen):
-    """Startup screen for the paper trading workspace."""
-
-    DEFAULT_CSS = """
-    ModeSelectScreen {
-        align: center middle;
-        background: #060606;
-    }
-    #mode-box {
-        width: 64;
-        height: auto;
-        max-height: 20;
-        background: #0a0a0a;
-        border: none;
-        padding: 0;
-    }
-    #mode-header {
-        height: 4; width: 100%; padding: 1 2;
-        background: #0a0a0a;
-    }
-    .mode-brand {
-        width: 100%; text-align: center;
-        color: #888888; height: 1;
-    }
-    .mode-version {
-        width: 100%; text-align: center;
-        color: #5d6670; height: 1;
-    }
-    #mode-options {
-        height: auto; width: 100%;
-        padding: 1 3;
-        background: #0a0a0a;
-    }
-    .mode-row {
-        width: 100%; height: 1;
-        padding: 0 2;
-        margin: 0 0;
-    }
-    #mode-footer {
-        height: 3; width: 100%;
-        padding: 1 2;
-        background: #0a0a0a;
-    }
-    .mode-keys {
-        width: 100%; text-align: center;
-        height: auto;
-    }
-    """
-
-    _selected: int = 0
-    _options = [
-        ("paper", "PAPER", "Paper trading workspace and local portfolio"),
-    ]
-
-    def compose(self) -> ComposeResult:
-        with Vertical(id="mode-box"):
-            with Vertical(id="mode-header"):
-                yield Static(
-                    Text.from_markup(
-                        "[bold #ffaf00]N E P S E[/]  [bold #e8e8e8]Q U A N T[/]"
-                    ),
-                    classes="mode-brand",
-                )
-                yield Static(
-                    Text.from_markup("[#5d6670]Terminal v3  ◆  Paper Trading[/]"),
-                    classes="mode-version",
-                )
-            with Vertical(id="mode-options"):
-                yield Static("", id="mode-opt-0", classes="mode-row")
-            with Vertical(id="mode-footer"):
-                yield Static(
-                    Text.from_markup(
-                        "[bold #ffaf00]\u2191\u2193[/][#555555] select   [/]"
-                        "[bold #ffaf00]ENTER[/][#555555] confirm   [/]"
-                        "[bold #ffaf00]Q[/][#555555] quit[/]"
-                    ),
-                    classes="mode-keys",
-                )
-
-    def on_mount(self) -> None:
-        self._render_options()
-
-    def _render_options(self) -> None:
-        for i, (key, label, desc) in enumerate(self._options):
-            widget = self.query_one(f"#mode-opt-{i}", Static)
-            if i == self._selected:
-                t = Text()
-                t.append("  \u25b6 ", style="bold #ffaf00")
-                t.append(f"{label:<10}", style="bold #e8e8e8")
-                t.append(desc, style="#888888")
-                widget.update(t)
-            else:
-                t = Text()
-                t.append("    ", style="")
-                t.append(f"{label:<10}", style="#555555")
-                t.append(desc, style="#333333")
-                widget.update(t)
-
-    def key_up(self) -> None:
-        self._selected = max(0, self._selected - 1)
-        self._render_options()
-
-    def key_down(self) -> None:
-        self._selected = min(len(self._options) - 1, self._selected + 1)
-        self._render_options()
-
-    def key_enter(self) -> None:
-        self.dismiss(self._options[self._selected][0])
-
-    def key_q(self) -> None:
-        self.app.exit()
-
-    def key_escape(self) -> None:
-        self.app.exit()
-
-
-class TMSLoginScreen(ModalScreen):
-    """Credential entry screen for live TMS auto-login."""
-
-    DEFAULT_CSS = """
-    TMSLoginScreen {
-        align: center middle;
-        background: #060606;
-    }
-    #tms-box {
-        width: 56;
-        height: auto;
-        max-height: 18;
-        background: #0a0a0a;
-        border: tall #333333;
-        padding: 1 2;
-    }
-    .tms-title {
-        width: 100%; text-align: center;
-        color: #ffaf00; text-style: bold;
-        height: 1; margin-bottom: 1;
-    }
-    .tms-label {
-        width: 100%; height: 1;
-        color: #666666; padding: 0 1;
-    }
-    .tms-input {
-        width: 100%;
-        background: #111111;
-        border: none;
-        border-bottom: solid #222222;
-        color: #e8e8e8;
-        margin-bottom: 1;
-    }
-    .tms-input:focus {
-        border-bottom: solid #ffaf00;
-    }
-    .tms-status {
-        width: 100%; text-align: center;
-        height: 1; color: #666666;
-        margin-top: 1;
-    }
-    .tms-hint {
-        width: 100%; text-align: center;
-        height: 1; margin-top: 1;
-    }
-    """
-
-    def compose(self) -> ComposeResult:
-        with Vertical(id="tms-box"):
-            yield Static(
-                Text.from_markup("[bold #ffaf00]TMS19 LOGIN[/]"),
-                classes="tms-title",
-            )
-            yield Static("Client ID / Username", classes="tms-label")
-            yield Input(id="tms-user", placeholder="e.g. 12345678", classes="tms-input")
-            yield Static("Password", classes="tms-label")
-            yield Input(id="tms-pass", placeholder="", password=True, classes="tms-input")
-            yield Static("", id="tms-status", classes="tms-status")
-            yield Static(
-                Text.from_markup(
-                    "[bold #ffaf00]ENTER[/][#555555] login   [/]"
-                    "[bold #ffaf00]ESC[/][#555555] back[/]"
-                ),
-                classes="tms-hint",
-            )
-
-    def on_mount(self) -> None:
-
-        settings = load_tms_settings()
-        user_env = settings.username or ""
-        pass_env = settings.password or ""
-        if user_env:
-            self.query_one("#tms-user", Input).value = user_env
-        if pass_env:
-            self.query_one("#tms-pass", Input).value = pass_env
-        if user_env and pass_env:
-            # Both pre-filled, focus submit hint
-            self.query_one("#tms-pass", Input).focus()
-        else:
-            self.query_one("#tms-user", Input).focus()
-
-    def on_input_submitted(self, event: Input.Submitted) -> None:
-        if event.input.id == "tms-user":
-            self.query_one("#tms-pass", Input).focus()
-        elif event.input.id == "tms-pass":
-            self._do_login()
-
-    def key_escape(self) -> None:
-        self.dismiss(None)
-
-    def _do_login(self) -> None:
-        user = self.query_one("#tms-user", Input).value.strip()
-        pwd = self.query_one("#tms-pass", Input).value.strip()
-        if not user or not pwd:
-            self.query_one("#tms-status", Static).update(
-                Text("Enter both client ID and password", style="#ff4444")
-            )
-            return
-        self.query_one("#tms-status", Static).update(
-            Text("Connecting to TMS19...", style="#ffaf00")
-        )
-        self.dismiss({"username": user, "password": pwd})
-
-
-# ═══════════════════════════════════════════════════════════════════════════════
-# COMMAND PALETTE (Bloomberg GO bar)
-# ═══════════════════════════════════════════════════════════════════════════════
-
-class CommandPalette(ModalScreen[dict | None]):
-    """Bloomberg-style GO bar. Type symbol name → lookup, tab name → switch, action → execute."""
-
-    DEFAULT_CSS = """
-    CommandPalette {
-        align: center top;
-        background: rgba(0, 0, 0, 0.80);
-    }
-    #cmd-box {
-        width: 64;
-        height: auto;
-        max-height: 22;
-        background: #090d12;
-        border: solid #3f474f;
-        margin-top: 2;
-        padding: 0;
-    }
-    #cmd-header {
-        height: 1;
-        width: 100%;
-        background: #090d12;
-        layout: horizontal;
-        padding: 0 1;
-    }
-    #cmd-header-label {
-        width: auto;
-        height: 1;
-        color: #f2b94b;
-        text-style: bold;
-        padding: 0 1 0 0;
-    }
-    #cmd-input {
-        width: 1fr;
-        background: #0c1116;
-        border: none;
-        color: #e8edf3;
-        padding: 0 1;
-        height: 1;
-    }
-    #cmd-input:focus {
-        background: #0c1116;
-        color: #fff6de;
-        border: none;
-    }
-    #cmd-results {
-        height: auto;
-        max-height: 18;
-        width: 100%;
-        background: #090d12;
-        padding: 0;
-        border-top: solid #1a1f25;
-    }
-    .cmd-row {
-        height: 1;
-        width: 100%;
-        padding: 0 1;
-        color: #6b7785;
-    }
-    .cmd-row-selected {
-        height: 1;
-        width: 100%;
-        padding: 0 1;
-        background: #111820;
-        color: #f2b94b;
-    }
-    #cmd-hint {
-        height: 1;
-        width: 100%;
-        background: #090d12;
-        color: #3a4450;
-        padding: 0 1;
-        border-top: solid #1a1f25;
-    }
-    """
-
-    _selected: int = 0
-    _items: list[dict] = []
-    _filtered: list[dict] = []
-
-    def compose(self) -> ComposeResult:
-        with Vertical(id="cmd-box"):
-            with Horizontal(id="cmd-header"):
-                yield Static("GO", id="cmd-header-label")
-                yield Input(id="cmd-input", placeholder="symbol, tab, or action")
-            yield Vertical(id="cmd-results")
-            yield Static(" ↑↓ Navigate  ENTER Execute  ESC Close", id="cmd-hint")
-
-    def on_mount(self) -> None:
-        self._build_items()
-        self._selected = 0
-        self._filter("")
-        self.query_one("#cmd-input", Input).focus()
-
-    def _build_items(self) -> None:
-        self._items = []
-        # Tab navigation
-        tabs = [
-            ("1 MARKET", "tab", "market", "Market overview — gainers, losers, volume"),
-            ("2 PORTFOLIO", "tab", "portfolio", "Portfolio holdings, NAV, risk"),
-            ("3 SIGNALS", "tab", "signals", "Trading signals, screener & calendar"),
-            ("4 LOOKUP", "tab", "lookup", "Stock lookup & charts"),
-            ("5 AGENTS", "tab", "agents", "AI agent analysis & chat"),
-            ("6 ORDERS", "tab", "orders", "Order management"),
-            ("7 WATCHLIST", "tab", "watchlist", "Your watched stocks"),
-            ("8 RATES & COMMODITIES", "tab", "kalimati", "FX, metals and local commodity prices"),
-            ("9 ACCOUNT", "tab", "account", "Paper account profiles and runtime"),
-            ("0 STRATEGIES", "tab", "strategies", "Saved strategies, bindings and backtests"),
-        ]
-        for name, kind, target, desc in tabs:
-            self._items.append({"name": name, "kind": kind, "target": target, "desc": desc})
-        # Actions
-        actions = [
-            ("BUY", "action", "buy", "Open buy dialog"),
-            ("SELL", "action", "sell", "Open sell dialog"),
-            ("REFRESH", "action", "refresh", "Refresh all market data"),
-            ("AGENT", "action", "agent", "Run AI agent analysis"),
-        ]
-        for name, kind, target, desc in actions:
-            self._items.append({"name": name, "kind": kind, "target": target, "desc": desc})
-        # Stock symbols from market data (will be populated dynamically)
-        try:
-            conn = _db()
-            import sqlite3
-            syms = [r[0] for r in conn.execute(
-                "SELECT DISTINCT symbol FROM stock_prices WHERE symbol != 'NEPSE' "
-                "ORDER BY symbol").fetchall()]
-            conn.close()
-            for sym in syms:
-                self._items.append({"name": sym, "kind": "stock", "target": sym,
-                                    "desc": "Look up stock"})
-        except Exception:
-            pass
-
-    def _filter(self, query: str) -> None:
-        q = query.strip().upper()
-        if not q:
-            # Show tabs and actions first
-            self._filtered = [i for i in self._items if i["kind"] != "stock"][:15]
-        else:
-            # Fuzzy match: items starting with query first, then contains
-            starts = [i for i in self._items if i["name"].upper().startswith(q)]
-            contains = [i for i in self._items if q in i["name"].upper() and i not in starts]
-            desc_match = [i for i in self._items if q in i["desc"].upper()
-                          and i not in starts and i not in contains]
-            self._filtered = (starts + contains + desc_match)[:15]
-        self._selected = min(self._selected, max(0, len(self._filtered) - 1))
-        self._render_results()
-
-    def _render_results(self) -> None:
-        container = self.query_one("#cmd-results", Vertical)
-        container.remove_children()
-        for i, item in enumerate(self._filtered):
-            kind_icon = {"tab": "◧", "action": "▶", "stock": "◆"}.get(item["kind"], "·")
-            kind_color = {"tab": CYAN, "action": GAIN_HI, "stock": AMBER}.get(item["kind"], LABEL)
-            t = Text()
-            if i == self._selected:
-                t.append(f"  ▸ {kind_icon} ", style=f"bold {kind_color}")
-                t.append(f"{item['name']:<16}", style=f"bold {WHITE}")
-                t.append(item["desc"], style=LABEL)
-                cls = "cmd-row-selected"
-            else:
-                t.append(f"    {kind_icon} ", style=kind_color)
-                t.append(f"{item['name']:<16}", style=DIM)
-                t.append(item["desc"], style=DIM)
-                cls = "cmd-row"
-            container.mount(Static(t, classes=cls))
-
-    def on_input_changed(self, event: Input.Changed) -> None:
-        if event.input.id == "cmd-input":
-            self._selected = 0
-            self._filter(event.value)
-
-    def on_input_submitted(self, event: Input.Submitted) -> None:
-        if event.input.id == "cmd-input":
-            self._execute()
-
-    def key_up(self) -> None:
-        self._selected = max(0, self._selected - 1)
-        self._render_results()
-
-    def key_down(self) -> None:
-        self._selected = min(len(self._filtered) - 1, self._selected + 1)
-        self._render_results()
-
-    def key_escape(self) -> None:
-        self.dismiss(None)
-
-    def _execute(self) -> None:
-        if not self._filtered:
-            # Treat input as stock symbol
-            val = self.query_one("#cmd-input", Input).value.strip().upper()
-            if val:
-                self.dismiss({"kind": "stock", "target": val})
-            return
-        item = self._filtered[self._selected]
-        self.dismiss(item)
-
-
-class WatchlistAddScreen(ModalScreen[dict | str | None]):
-    """Bloomberg-style watchlist picker with live suggestions."""
-
-    DEFAULT_CSS = """
-    WatchlistAddScreen {
-        align: center middle;
-        background: rgba(0, 0, 0, 0.88);
-    }
-    #wl-add-box {
-        width: 76;
-        height: auto;
-        max-height: 24;
-        border: solid #5d6670;
-        background: #090b0e;
-        padding: 0;
-    }
-    #wl-add-title {
-        height: 1;
-        width: 100%;
-        background: #17191d;
-        color: #ffaf00;
-        text-style: bold;
-        padding: 0 2;
-    }
-    #wl-add-row {
-        width: 100%;
-        height: 3;
-        background: #101419;
-        border-top: solid #232a31;
-        border-bottom: solid #232a31;
-        padding: 0 2;
-        layout: horizontal;
-        content-align: left middle;
-    }
-    #wl-add-label {
-        height: 1;
-        width: 8;
-        padding-right: 2;
-        color: #8fa0b1;
-        text-style: bold;
-        content-align: left middle;
-    }
-    #wl-add-input {
-        width: 1fr;
-        height: 1;
-        background: #101419;
-        border: none;
-        color: #f6fbff;
-        padding: 0 1;
-        content-align: left middle;
-    }
-    #wl-add-input:focus {
-        background: #151a20;
-        color: #ffffff;
-        border: none;
-    }
-    #wl-add-query {
-        height: 1;
-        width: 100%;
-        background: #0d1014;
-        color: #8fa0b1;
-        padding: 0 2;
-        border-bottom: solid #1b2128;
-    }
-    #wl-add-results {
-        height: auto;
-        max-height: 17;
-        width: 100%;
-        background: #0b0d10;
-        padding: 1 0;
-    }
-    .wl-add-row-item {
-        height: 1;
-        width: 100%;
-        padding: 0 2;
-        color: #7e8791;
-    }
-    .wl-add-row-selected {
-        height: 1;
-        width: 100%;
-        padding: 0 2;
-        background: #191d22;
-        color: #ffcf70;
-        text-style: bold;
-    }
-    #wl-add-hint {
-        height: 1;
-        width: 100%;
-        background: #101318;
-        color: #56616d;
-        padding: 0 2;
-        border-top: solid #1b2128;
-    }
-    """
-
-    _selected: int = 0
-    _items: list[dict] = []
-    _filtered: list[dict] = []
-
-    def compose(self) -> ComposeResult:
-        with Vertical(id="wl-add-box"):
-            yield Static("ADD TO WATCHLIST", id="wl-add-title")
-            with Horizontal(id="wl-add-row"):
-                yield Static("SEARCH", id="wl-add-label")
-                yield Input(id="wl-add-input", placeholder="NABIL, USD, Gold / Tola, Petrol...")
-            yield Static("", id="wl-add-query")
-            yield Vertical(id="wl-add-results")
-            yield Static("↑↓ Navigate   ENTER Add   ESC Cancel   Stocks, forex, macro, commodities", id="wl-add-hint")
-
-    def on_mount(self) -> None:
-        self._build_items()
-        self._selected = 0
-        self._filter("")
-        self._update_query_preview("")
-        self.query_one("#wl-add-input", Input).focus()
-
-    def _build_items(self) -> None:
-        self._items = []
-        try:
-            conn = _db()
-            symbols = [r[0] for r in conn.execute(
-                "SELECT DISTINCT symbol FROM stock_prices WHERE symbol != 'NEPSE' ORDER BY symbol"
-            ).fetchall()]
-            conn.close()
-            for sym in symbols:
-                self._items.append({
-                    "name": sym,
-                    "kind": "stock",
-                    "target": _stock_watchlist_entry(sym),
-                    "desc": "NEPSE stock",
-                })
-        except Exception:
-            pass
-
-        app = self.app
-        kalimati_rows = list(getattr(app, "_kalimati_rows", []) or [])
-        macro_rates = dict(getattr(app, "_macro_rates", {}) or {})
-        if not kalimati_rows:
-            try:
-                kalimati_rows = list(get_kalimati_display_rows() or [])
-            except Exception:
-                kalimati_rows = []
-        if not macro_rates:
-            try:
-                metals = _fetch_gold_silver_prices()
-                indicators: list[dict] = []
-                if metals:
-                    gold = float(metals.get("gold_per_tola") or 0)
-                    silver = float(metals.get("silver_per_tola") or 0)
-                    if gold > 0:
-                        indicators.append({"item": "Gold / Tola", "group": "Metals", "unit": "NPR/tola"})
-                    if silver > 0:
-                        indicators.append({"item": "Silver / Tola", "group": "Metals", "unit": "NPR/tola"})
-                forex_rows = _fetch_nrb_forex_rates(("USD", "EUR", "GBP", "INR", "CNY", "JPY"))
-                macro_rates = {"indicators": indicators, "forex_rows": forex_rows}
-            except Exception:
-                macro_rates = {}
-
-        for row in kalimati_rows:
-            label = str(row.get("name_english") or "").strip()
-            if not label:
-                continue
-            self._items.append({
-                "name": label,
-                "kind": "commodity",
-                "target": {
-                    "kind": "commodity",
-                    "key": f"commodity:{label}",
-                    "label": label,
-                    "unit": str(row.get("unit") or ""),
-                },
-                "desc": f"Kalimati commodity  {row.get('unit') or ''}".strip(),
-            })
-
-        for row in list(macro_rates.get("indicators", []) or []):
-            label = str(row.get("item") or "").strip()
-            if not label:
-                continue
-            self._items.append({
-                "name": label,
-                "kind": "macro",
-                "target": {
-                    "kind": "macro",
-                    "key": f"macro:{label}",
-                    "label": label,
-                    "group": str(row.get("group") or ""),
-                },
-                "desc": f"{row.get('group') or 'Macro'}  {row.get('unit') or ''}".strip(),
-            })
-
-        for row in list(macro_rates.get("forex_rows", []) or []):
-            code = str(row.get("currency_code") or "").strip().upper()
-            name = str(row.get("currency_name") or "").strip()
-            if not code:
-                continue
-            self._items.append({
-                "name": code,
-                "kind": "forex",
-                "target": {
-                    "kind": "forex",
-                    "key": f"forex:{code}",
-                    "label": code,
-                    "currency_name": name,
-                },
-                "desc": name or "NRB forex rate",
-            })
-
-    def _filter(self, query: str) -> None:
-        q = query.strip().upper()
-        if not q:
-            priority = ("stock", "forex", "macro", "commodity")
-            ordered: list[dict] = []
-            for kind in priority:
-                ordered.extend([item for item in self._items if item["kind"] == kind][:4])
-            self._filtered = ordered[:12]
-        else:
-            exact = [item for item in self._items if item["name"].upper() == q]
-            starts = [item for item in self._items if item["name"].upper().startswith(q) and item not in exact]
-            contains = [item for item in self._items if q in item["name"].upper() and item not in exact and item not in starts]
-            desc = [item for item in self._items if q in item["desc"].upper() and item not in exact and item not in starts and item not in contains]
-            self._filtered = (exact + starts + contains + desc)[:12]
-        self._selected = min(self._selected, max(0, len(self._filtered) - 1))
-        self._render_results()
-
-    def _render_results(self) -> None:
-        container = self.query_one("#wl-add-results", Vertical)
-        container.remove_children()
-        if not self._filtered:
-            container.mount(Static(Text("  No matches", style=DIM), classes="wl-add-row-item"))
-            return
-        kind_icon = {"stock": "◆", "forex": "FX", "macro": "●", "commodity": "◧"}
-        kind_style = {"stock": AMBER, "forex": CYAN, "macro": GAIN_HI, "commodity": WHITE}
-        for idx, item in enumerate(self._filtered):
-            t = Text()
-            icon = kind_icon.get(item["kind"], "·")
-            color = kind_style.get(item["kind"], LABEL)
-            if idx == self._selected:
-                t.append(f"  ▸ {icon} ", style=f"bold {color}")
-                t.append(f"{item['name']:<20}", style=f"bold {WHITE}")
-                t.append(item["desc"][:42], style=LABEL)
-                cls = "wl-add-row-selected"
-            else:
-                t.append(f"    {icon} ", style=color)
-                t.append(f"{item['name']:<20}", style=DIM)
-                t.append(item["desc"][:42], style=DIM)
-                cls = "wl-add-row-item"
-            container.mount(Static(t, classes=cls))
-
-    def _update_query_preview(self, query: str) -> None:
-        widget = self.query_one("#wl-add-query", Static)
-        clean = query.strip()
-        if clean:
-            widget.update(Text.from_markup(
-                f"[#6e7c89]QUERY[/] [bold {WHITE}]{_escape_markup(clean)}[/]   "
-                f"[#6e7c89]MATCHES[/] [bold {AMBER}]{len(self._filtered)}[/]"
-            ))
-        else:
-            widget.update(Text.from_markup(
-                f"[#6e7c89]QUERY[/] [#55616d]Start typing to search stocks, FX, macro and commodities[/]"
-            ))
-
-    def on_input_changed(self, event: Input.Changed) -> None:
-        if event.input.id == "wl-add-input":
-            self._selected = 0
-            self._filter(event.value)
-            self._update_query_preview(event.value)
-
-    def on_input_submitted(self, event: Input.Submitted) -> None:
-        if event.input.id == "wl-add-input":
-            self._submit()
-
-    def key_up(self) -> None:
-        self._selected = max(0, self._selected - 1)
-        self._render_results()
-
-    def key_down(self) -> None:
-        self._selected = min(len(self._filtered) - 1, self._selected + 1)
-        self._render_results()
-
-    def _submit(self) -> None:
-        raw = self.query_one("#wl-add-input", Input).value.strip()
-        if self._filtered:
-            self.dismiss(self._filtered[self._selected]["target"])
-            return
-        if raw:
-            self.dismiss(raw.upper())
-        else:
-            self.dismiss(None)
-
-    def key_escape(self) -> None:
-        self.dismiss(None)
-
-
-class LookupScreen(ModalScreen[dict | None]):
-    """Compact symbol lookup prompt."""
-
-    DEFAULT_CSS = """
-    LookupScreen {
-        align: center middle;
-        background: rgba(0, 0, 0, 0.82);
-    }
-    #lookup-box {
-        width: 38;
-        height: 3;
-        border: solid #3f474f;
-        background: #090d12;
-        padding: 0 1;
-        layout: horizontal;
-        align: left middle;
-    }
-    #lookup-label {
-        color: #f2b94b;
-        text-style: bold;
-        height: 1;
-        width: auto;
-        padding: 0 1 0 0;
-    }
-    #lookup-input {
-        width: 1fr;
-        height: 1;
-        background: #0c1116;
-        border: none;
-        color: #e8edf3;
-        padding: 0 1;
-    }
-    #lookup-input:focus {
-        background: #0c1116;
-        color: #fff6de;
-        border: none;
-    }
-    #lookup-hint {
-        height: 1;
-        width: auto;
-        color: #4a5562;
-        padding: 0 0 0 1;
-    }
-    """
-
-    def compose(self) -> ComposeResult:
-        with Horizontal(id="lookup-box"):
-            yield Static("LOOKUP", id="lookup-label")
-            yield Input(id="lookup-input", placeholder="NABIL")
-            yield Static("↵ GO  ESC ✕", id="lookup-hint")
-
-    def on_mount(self) -> None:
-        self.query_one("#lookup-input", Input).focus()
-
-    def on_input_submitted(self, event: Input.Submitted) -> None:
-        self._submit()
-
-    def _submit(self) -> None:
-        val = self.query_one("#lookup-input", Input).value.strip().upper()
-        if val:
-            self.dismiss({"symbol": val})
-        else:
-            self.dismiss(None)
-
-    def key_escape(self) -> None:
-        self.dismiss(None)
+from apps.tui.screens.mode_select import ModeSelectScreen  # re-export
+from apps.tui.screens.tms_login import TMSLoginScreen, load_tms_settings  # re-export
+from apps.tui.screens.command_palette import CommandPalette  # re-export
+from apps.tui.screens.watchlist_add import WatchlistAddScreen  # re-export
+from apps.tui.screens.lookup import LookupScreen  # re-export
 
 
 # ═══════════════════════════════════════════════════════════════════════════════
